+++ sources/recordPlayer2fridge.cpp
@@ -26,7 +26,7 @@ int8_t netOpt::recordPlayer2fridge(roomMember *recordPlayer, roomMember *fridge)
         {
             cout << "." << (int)mac[i];
         }
-        cout << dec << " with probability adjustment of " << getProbAdjustment(d1, OV1, 0.0) << endl;
+        cout << dec << " with probability adjustment of " << (int)getProbAdjustment(d1, OV1, 0.0) << endl;
     #endif
 
     if(d1->activity.getLen() < 2 || OV1->activity.getLen() < 2)
