+++ sources/speaker2washing.cpp
@@ -26,7 +26,7 @@ int8_t netOpt::speaker2washing(roomMember *speaker, roomMember *washing)
         {
             cout << "." << (int)mac[i];
         }
-        cout << dec << " with probability adjustment of " << getProbAdjustment(((devRecord *)g1->mems.getHead()->data), d2, 0.0) << endl;
+        cout << dec << " with probability adjustment of " << (int)getProbAdjustment(((devRecord *)g1->mems.getHead()->data), d2, 0.0) << endl;
     #endif
 
     if(((devRecord *)g1->mems.getHead()->data)->activity.getLen() < 2 || d2->activity.getLen() < 2)
