+++ sources/tv2speaker.cpp
@@ -26,7 +26,7 @@ int8_t netOpt::tv2speaker(roomMember *tv, roomMember *speaker)
         {
             cout << "." << (int)mac[i];
         }
-        cout << dec << " with probability adjustment of " << getProbAdjustment(d1, ((devRecord *)g1->mems.getHead()->data), 0.0) << endl;
+        cout << dec << " with probability adjustment of " << (int)getProbAdjustment(d1, ((devRecord *)g1->mems.getHead()->data), 0.0) << endl;
     #endif
 
     if(d1->activity.getLen() < 2 || ((devRecord *)g1->mems.getHead()->data)->activity.getLen() < 2)
